pi/requirements.txt
fastapi
uvicorn
paho-mqtt
sqlite-utils

pytest
httpx
pytest-asyncio
